+++ backend/apid/routers/entities.go
@@ -7,6 +7,7 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/sensu/sensu-go/backend/apid/actions"
 	"github.com/sensu/sensu-go/backend/store"
+	"github.com/sensu/sensu-go/types"
 )
 
 // EntitiesRouter handles requests for /entities
@@ -27,6 +28,7 @@ func (r *EntitiesRouter) Mount(parent *mux.Router) {
 	routes.getAll(r.list)
 	routes.get(r.find)
 	routes.del(r.destroy)
+	routes.post(r.create)
 }
 
 func (r *EntitiesRouter) destroy(req *http.Request) (interface{}, error) {
@@ -52,4 +54,13 @@ func (r *EntitiesRouter) find(req *http.Request) (interface{}, error) {
 func (r *EntitiesRouter) list(req *http.Request) (interface{}, error) {
 	records, err := r.controller.Query(req.Context())
 	return records, err
+}
+
+func (r *EntitiesRouter) create(req *http.Request) (interface{}, error) {
+	entity := types.Entity{}
+	if err := unmarshalBody(req, &entity); err != nil {
+		return nil, err
+	}
+	err := r.controller.Create(req.Context(), entity)
+	return entity, err
 }